# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/perizinan": {
            "get": {
                "tags": [
                    "Perizinan"
                ],
                "summary": "Get all perizinan requests",
                "operationId": "getPerizinan",
                "responses": {
                    "200": {
                        "description": "Returns all perizinan requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Perizinan"
                ],
                "summary": "Create a new perizinan request",
                "operationId": "createPerizinan",
                "requestBody": {
                    "description": "Request body for creating a new perizinan request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRequestPerizinan"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Perizinan request created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/perizinan/{id}": {
            "get": {
                "tags": [
                    "Perizinan"
                ],
                "summary": "Get a perizinan request by ID",
                "operationId": "getPerizinanById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the perizinan request to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the specified perizinan request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Perizinan request not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Perizinan"
                ],
                "summary": "Update a perizinan request by ID",
                "operationId": "updatePerizinan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the perizinan request to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request body for updating a perizinan request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRequestPerizinan"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perizinan request updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Perizinan request not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Perizinan"
                ],
                "summary": "Delete a perizinan request by ID",
                "operationId": "deletePerizinan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the perizinan request to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Perizinan request deleted successfully"
                    },
                    "404": {
                        "description": "Perizinan request not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "DataResponse": {
                "description": "Data response dari permintaan perizinan",
                "properties": {
                    "status": {
                        "description": "Keterangan status response",
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "description": "Pesan terkait response",
                        "type": "string",
                        "example": "Permintaan perizinan berhasil diajukan"
                    },
                    "data": {
                        "description": "Data API",
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "status": "success",
                    "message": "Permintaan perizinan berhasil diajukan",
                    "data": {
                        "nama": "John Doe",
                        "alamat": "Jl. Kenanga No. 123",
                        "jenis_perizinan": "Izin Usaha",
                        "tanggal_mulai": "2023-04-01",
                        "tanggal_selesai": "2023-06-30"
                    }
                }
            },
...
            "NotFound": {
                "description": "Data yang diminta tidak ditemukan",
                "properties": {
                    "message": {
                        "description": "Pesan error yang menunjukkan data tidak ditemukan",
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "message": "Data dengan ID 123 tidak ditemukan"
                }
            },
            "CreateRequestPerizinan": {
                "title": "Create Request Perizinan",
                "required": [
                    "nama",
                    "alamat",
                    "warga_id"
                ],
                "properties": {
                    "nama": {
                        "description": "Nama lengkap pemohon perizinan",
                        "type": "string"
                    },
                    "alamat": {
                        "description": "Alamat lengkap pemohon perizinan",
                        "type": "string"
                    },
                    "tanggal_request": {
                        "description": "Tanggal request perizinan",
                        "type": "string"
                    },
                    "keterangan": {
                        "description": "Keterangan perizinan",
                        "type": "string"
                    },
                    "lampiran": {
                        "description": "Lampiran perizinan",
                        "type": "string"
                    },
                    "status_request": {
                        "description": "Status perizinan",
                        "type": "string"
                    },
                    "jenis_perizinan": {
                        "description": "Jenis perizinan yang diajukan",
                        "type": "string"
                    },
                    "jenis_id": {
                        "description": "ID dari perizinan yang diajukan",
                        "type": "integer"
                    },
                    "warga_id": {
                        "description": "ID dari warga yang mengajukan perizinan",
                        "type": "integer"
                    }
                },
                "type": "object",
                "example": {
                    "nama": "John Doe",
                    "alamat": "Jl. Sudirman No. 123",
                    "jenis_perizinan": "Surat Izin Keramaian",
                    "tanggal_request": "2022-01-01",
                    "status_request": "",
                    "keterangan": "Pembuatan surat izin keramaian",
                    "warga_id": 1
                }
            },
            "UpdateRequestPerizinan": {
                "title": "Update Request Perizinan",
                "description": "Data yang diperlukan untuk memperbarui permintaan perizinan",
                "required": [
                    "nama",
                    "alamat",
                    "jenis_perizinan",
                    "jenis_id",
                    "tanggal_mulai",
                    "tanggal_selesai"
                ],
                "properties": {
                    "nama": {
                        "description": "Nama lengkap pemohon perizinan",
                        "type": "string"
                    },
                    "alamat": {
                        "description": "Alamat lengkap pemohon perizinan",
                        "type": "string"
                    },
                    "jenis_perizinan": {
                        "description": "Jenis perizinan yang diminta",
                        "type": "string"
                    },
                    "jenis_id": {
                        "description": "ID dari perizinan yang diajukan",
                        "type": "integer"
                    },
                    "tanggal_request": {
                        "description": "Tanggal request perizinan",
                        "type": "string"
                    },
                    "keterangan": {
                        "description": "Keterangan perizinan",
                        "type": "string"
                    },
                    "status_request": {
                        "description": "Status perizinan",
                        "type": "string"
                    },
                    "tanggal_mulai": {
                        "description": "Tanggal mulai berlakunya perizinan",
                        "type": "string",
                        "format": "date"
                    },
                    "tanggal_selesai": {
                        "description": "Tanggal berakhirnya perizinan",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object",
                "example": {
                    "nama": "John Doe",
                    "alamat": "Jl. Kenanga No. 123",
                    "jenis_perizinan": "Izin Usaha",
                    "tanggal_request": "2022-01-01",
                    "status_request": "",
                    "keterangan": "Pembuatan surat izin usaha",
                    "jenis_id": 123,
                    "tanggal_mulai": "2023-04-01",
                    "tanggal_selesai": "2023-06-30"
                }
            },
            "ValidationError": {
                "description": "Error validasi data request",
                "properties": {
                    "message": {
                        "description": "Pesan error yang menunjukkan data request tidak valid",
                        "type": "string"
                    },
                    "errors": {
                        "description": "Daftar error validasi pada setiap properti request",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object",
                "example": {
                    "message": "Data request tidak valid",
                    "errors": [
                        "Nama harus diisi",
                        "Alamat harus diisi",
                        "Jenis perizinan harus diisi",
                        "Warga ID harus diisi"
                    ]
                }
            },
...
        }
    }
}